code improvements
SDK_INT is never <9 because minSdkVersion 9
StringBuffer replaced by StringBuilder
Explicit manual boxing is unnecessary under Java 5 and newer
`while` loops replaced by foreach
`for i` replaced by foreach
Inlined some variables
Missing supercall fix

Closes #664

+++ src/org/smssecure/smssecure/contacts/ContactAccessor.java
@@ -258,9 +258,9 @@ public class ContactAccessor {
 
     if (phone.length() > 0) {
       ArrayList result = new ArrayList();
-      result.add(Integer.valueOf(-1));                    // ID
-      result.add(Long.valueOf(-1));                       // CONTACT_ID
-      result.add(Integer.valueOf(Phone.TYPE_CUSTOM));     // TYPE
+      result.add(-1);                    // ID
+      result.add((long) -1);             // CONTACT_ID
+      result.add(Phone.TYPE_CUSTOM);     // TYPE
       result.add(phone);                                  // NUMBER
 
     /*
